requirements.txt
anndata>=0.10
beautifulsoup4>=4.12
lxml>=5.2
matplotlib>=3.8
networkx>=3.2
numpy>=1.26
openpyxl>=3.1
pandas>=2.2
psutil>=5.9
pyarrow>=16.0
pyyaml>=6.0
requests>=2.32
scanpy>=1.10
scikit-learn>=1.4
scipy>=1.12
seaborn>=0.13
torch>=2.2
tqdm>=4.66
